Move metadata validation to MetadataValidationService
